+++ server/templates/shaders/fragment/shared/pixel-distance/blue-distance.glsl
@@ -1,0 +1,3 @@
+float quick_distance(vec3 pixel1, vec3 pixel2){
+    return abs(pixel1.b - pixel2.b);
+}