+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+* Add red, green, and blue color comparisons for dithering
 * Add color palette
 
 ## 5.1 2026-01-09 [Commit 709ae55](https://github.com/allen-garvey/dithermark/commit/709ae553025dc8944baa436c54b76e3fe2689952)